requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
pygame==2.6.1
PyOpenGL==3.1.7
